fixed documentation and made everything uniform for pages
