+++ front-end/src/components/logo.png
