+++ front-end/src/pages/AISuggestions.css
