+++ front-end/src/pages/App.jsx
@@ -1,22 +1,22 @@
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
-import DailyTasks from "./pages/DailyTasks.jsx";
-import Placeholder from "./pages/Placeholder.jsx";
-import Settings from "./Settings.js";
-import AllTasks from "./pages/AllTasks.jsx";
+import DailyTasks from "./DailyTasks.jsx";
+import Placeholder from "./Placeholder.jsx";
+import Settings from "./Settings.jsx";
+import AllTasks from "./AllTasks.jsx";
 import CalendarPage from "./Calendar.jsx";
-import AddTask from "./pages/AddTasks.jsx";
-import ChangeBio from "./ChangeBio.js";
-import Profile from "./profilePage.js";
-import EditTask from "./pages/EditTask.jsx";
-import ChangeMajor from "./ChangeMajor.js";
-import ChangeSchool from "./ChangeSchool.js";
-import AiSuggestions from "./AISuggestions.js";
-import Home from "./homePage.js";
-import ChangeTimezone from "./ChangeTimezone.js";
-import Login from "./loginPage.js";
-import Goals from "./Goals.js";
-import Register from "./registerPage.js";
-import Dashboard from "./dashboardPage.js";
+import AddTask from "./AddTasks.jsx";
+import ChangeBio from "./ChangeBio.jsx";
+import Profile from "./profilePage.jsx";
+import EditTask from "./EditTask.jsx";
+import ChangeMajor from "./ChangeMajor.jsx";
+import ChangeSchool from "./ChangeSchool.jsx";
+import AiSuggestions from "./AISuggestions.jsx";
+import Home from "./homePage.jsx";
+import ChangeTimezone from "./ChangeTimezone.jsx";
+import Login from "./loginPage.jsx";
+import Goals from "./Goals.jsx";
+import Register from "./registerPage.jsx";
+import Dashboard from "./dashboardPage.jsx";
 
 const fmtLocalDate = (d) => {
   const y = d.getFullYear();
