+++ front-end/src/pages/ChangeBio.jsx
@@ -1,24 +1,24 @@
-import './ChangeMajor.css'
+import './ChangeBio.css'
 /*import logo from './logo.png'*/
-import HeaderBar from "./components/HeaderBar.jsx";
-import BottomNav from "./components/BottomNav.jsx";
+import HeaderBar from "../components/HeaderBar.jsx";
+import BottomNav from "../components/BottomNav.jsx";
 import { useState } from "react";
-import MenuOverlay from "./components/MenuOverlay.jsx";
+import MenuOverlay from "../components/MenuOverlay.jsx";
 import { useNavigate } from "react-router-dom";
 
-const ChangeMajor = () => {
+const ChangeBio = () => {
     const navigate = useNavigate();
     const [menuOpen, setMenuOpen] = useState(false);
     
     return(
-        <div className = "ChangeMajor">
-            <HeaderBar title="Change Major" onHamburger={() => setMenuOpen(true)} onLogo={() => {}} />
+        <div className = "ChangeBio">
+            <HeaderBar title="Change Bio" onHamburger={() => setMenuOpen(true)} onLogo={() => {}} />
 
-            <div className = "major-content">
+            <div className = "bio-content">
                 <h1>Settings</h1>
-                <h4>Major:</h4>
+                <h4>Bio:</h4>
                 <label>
-                    <input name="myInput" defaultValue="Major"/>
+                    <input name="myInput" defaultValue="Bio"/>
                 </label>
                 <button class="save-button" type="button">Save</button>
                 <button className="back-button" onClick={() => navigate("/settings")}>
@@ -32,4 +32,4 @@ const ChangeMajor = () => {
     );
 };
 
-export default ChangeMajor
+export default ChangeBio
