+++ front-end/src/pages/ChangeMajor.css
