+++ front-end/src/pages/ChangeSchool.jsx
@@ -1,24 +1,24 @@
-import './ChangeBio.css'
+import './ChangeSchool.css'
 /*import logo from './logo.png'*/
-import HeaderBar from "./components/HeaderBar.jsx";
-import BottomNav from "./components/BottomNav.jsx";
+import HeaderBar from "../components/HeaderBar.jsx";
+import BottomNav from "../components/BottomNav.jsx";
 import { useState } from "react";
-import MenuOverlay from "./components/MenuOverlay.jsx";
+import MenuOverlay from "../components/MenuOverlay.jsx";
 import { useNavigate } from "react-router-dom";
 
-const ChangeBio = () => {
+const ChangeSchool = () => {
     const navigate = useNavigate();
     const [menuOpen, setMenuOpen] = useState(false);
     
     return(
-        <div className = "ChangeBio">
-            <HeaderBar title="Change Bio" onHamburger={() => setMenuOpen(true)} onLogo={() => {}} />
+        <div className = "ChangeSchool">
+            <HeaderBar title="Change School" onHamburger={() => setMenuOpen(true)} onLogo={() => {}} />
 
-            <div className = "bio-content">
+            <div className = "school-content">
                 <h1>Settings</h1>
-                <h4>Bio:</h4>
+                <h4>School:</h4>
                 <label>
-                    <input name="myInput" defaultValue="Bio"/>
+                    <input name="myInput" defaultValue="School"/>
                 </label>
                 <button class="save-button" type="button">Save</button>
                 <button className="back-button" onClick={() => navigate("/settings")}>
@@ -32,4 +32,4 @@ const ChangeBio = () => {
     );
 };
 
-export default ChangeBio
+export default ChangeSchool
