+++ front-end/src/pages/Goals.jsx
@@ -1,9 +1,9 @@
 import './Goals.css'
 /*import logo from './logo.png'*/
-import HeaderBar from "./components/HeaderBar.jsx";
-import BottomNav from "./components/BottomNav.jsx";
+import HeaderBar from "../components/HeaderBar.jsx";
+import BottomNav from "../components/BottomNav.jsx";
 import { useState } from "react";
-import MenuOverlay from "./components/MenuOverlay.jsx";
+import MenuOverlay from "../components/MenuOverlay.jsx";
 import { useNavigate } from "react-router-dom";
 import EditGoal from "./EditGoal.jsx";
