+++ front-end/src/pages/loginPage.jsx
@@ -1,45 +1,36 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import HeaderBar from "./components/HeaderBar.jsx";
-import BottomNav from "./components/BottomNav.jsx";
-import MenuOverlay from "./components/MenuOverlay.jsx";
-import "./registerPage.css";
+import HeaderBar from "../components/HeaderBar.jsx";
+import BottomNav from "../components/BottomNav.jsx";
+import MenuOverlay from "../components/MenuOverlay.jsx";
+import "./loginPage.css";
 
-export default function RegisterPage() {
+export default function LoginPage() {
     const navigate = useNavigate();
     const [menuOpen, setMenuOpen] = useState(false);
     const [username, setUsername] = useState("");
     const [password, setPassword] = useState("");
-    const [name, setName] = useState("");
 
     const handleSubmit = (e) => {
         e.preventDefault();
-        console.log("Name:", name, "Username:", username, "Password:", password);
+        console.log("Username:", username, "Password:", password);
         navigate("/dashboard");
     };
 
     return (
-        <div className="page register-page">
+        <div className="page login-page">
         <HeaderBar 
-            title="Register"
+            title="Log In"
             onHamburger={() => setMenuOpen(true)}
             onLogo={() => navigate("/")} />
 
-        <main className="register-content">
+        <main className="login-content">
             <h1>Enter Details</h1>
-            <form className="register-form" onSubmit={handleSubmit}>
-            <input
-                type="text"
-                placeholder="Name"
-                className="register-input"
-                value={name}
-                onChange={(e) => setName(e.target.value)}
-                required
-            />
+            <form className="login-form" onSubmit={handleSubmit}>
             <input
                 type="text"
                 placeholder="Username"
-                className="register-input"
+                className="login-input"
                 value={username}
                 onChange={(e) => setUsername(e.target.value)}
                 required
@@ -47,13 +38,13 @@ export default function RegisterPage() {
             <input
                 type="password"
                 placeholder="Password"
-                className="register-input"
+                className="login-input"
                 value={password}
                 onChange={(e) => setPassword(e.target.value)}
                 required
             />
 
-            <div className="register-buttons">
+            <div className="login-buttons">
                 <button
                 type="button"
                 className="pixel-button"
@@ -68,6 +59,7 @@ export default function RegisterPage() {
             </form>
         </main>
 
+        {/* <BottomNav /> */}
         {menuOpen && <MenuOverlay onClose={() => setMenuOpen(false)} />}
         </div>
     );
