+++ front-end/src/pages/registerPage.jsx
@@ -1,36 +1,45 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import HeaderBar from "./components/HeaderBar.jsx";
-import BottomNav from "./components/BottomNav.jsx";
-import MenuOverlay from "./components/MenuOverlay.jsx";
-import "./loginPage.css";
+import HeaderBar from "../components/HeaderBar.jsx";
+import BottomNav from "../components/BottomNav.jsx";
+import MenuOverlay from "../components/MenuOverlay.jsx";
+import "./registerPage.css";
 
-export default function LoginPage() {
+export default function RegisterPage() {
     const navigate = useNavigate();
     const [menuOpen, setMenuOpen] = useState(false);
     const [username, setUsername] = useState("");
     const [password, setPassword] = useState("");
+    const [name, setName] = useState("");
 
     const handleSubmit = (e) => {
         e.preventDefault();
-        console.log("Username:", username, "Password:", password);
+        console.log("Name:", name, "Username:", username, "Password:", password);
         navigate("/dashboard");
     };
 
     return (
-        <div className="page login-page">
+        <div className="page register-page">
         <HeaderBar 
-            title="Log In"
+            title="Register"
             onHamburger={() => setMenuOpen(true)}
             onLogo={() => navigate("/")} />
 
-        <main className="login-content">
+        <main className="register-content">
             <h1>Enter Details</h1>
-            <form className="login-form" onSubmit={handleSubmit}>
+            <form className="register-form" onSubmit={handleSubmit}>
+            <input
+                type="text"
+                placeholder="Name"
+                className="register-input"
+                value={name}
+                onChange={(e) => setName(e.target.value)}
+                required
+            />
             <input
                 type="text"
                 placeholder="Username"
-                className="login-input"
+                className="register-input"
                 value={username}
                 onChange={(e) => setUsername(e.target.value)}
                 required
@@ -38,13 +47,13 @@ export default function LoginPage() {
             <input
                 type="password"
                 placeholder="Password"
-                className="login-input"
+                className="register-input"
                 value={password}
                 onChange={(e) => setPassword(e.target.value)}
                 required
             />
 
-            <div className="login-buttons">
+            <div className="register-buttons">
                 <button
                 type="button"
                 className="pixel-button"
@@ -59,7 +68,6 @@ export default function LoginPage() {
             </form>
         </main>
 
-        {/* <BottomNav /> */}
         {menuOpen && <MenuOverlay onClose={() => setMenuOpen(false)} />}
         </div>
     );
